Ajout du chemin Relatif pour ne plus avoir a choisir entre nos deux chemin absolue et resolution des arcs probleme de type

+++ src/Edge.java
@@ -37,7 +37,7 @@ public class Edge {
         return this.valeurs;
     }
 
-    public void setValeurs(int index, int valeur) {
+    public void setValeurs(int index, double valeur) {
         this.valeurs[index] = valeur;
     }
     //endregion
